# clients/shared/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "Fatoura API",
    "description": "UAE VAT sales, purchases and invoicing.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/quotations": {
      "get": {
        "tags": [
          "Quotations"
        ],
        "parameters": [
          {
            "name": "search",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "clientId",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "status",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/QuotationStatus"
            }
          },
          {
            "name": "from",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "to",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "openOnly",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PagedResultOfQuotationSummaryDto"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Quotations"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QuotationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuotationDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/quotations/{id}": {
      "get": {
        "tags": [
          "Quotations"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuotationDto"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Quotations"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QuotationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuotationDto"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Quotations"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        }
      }
    },
    "/api/quotations/{id}/status": {
      "put": {
        "tags": [
          "Quotations"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QuotationStatusRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuotationDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/quotations/{id}/convert": {
      "post": {
        "tags": [
          "Quotations"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "oneOf": [
                  {
                    "type": "null"
                  },
                  {
                    "$ref": "#/components/schemas/ConvertQuotationRequest"
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvoiceResult"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/invoices/{id}": {
      "get": {
        "tags": [
          "Invoices"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvoiceDto"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Invoices"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateInvoiceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvoiceResult"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ConvertQuotationRequest": {
        "required": [
          "payment"
        ],
        "type": "object",
        "properties": {
          "payment": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/PaymentRequest"
              }
            ]
          }
        }
      },
...
      "DiscountKind": {
        "enum": [
          "None",
          "Amount",
          "Percent"
        ]
      },
...
      "DocumentDiscountRequest": {
        "required": [
          "kind",
          "value"
        ],
        "type": "object",
        "properties": {
          "kind": {
            "$ref": "#/components/schemas/DiscountKind"
          },
          "value": {
            "type": "number",
            "format": "double"
          }
        },
        "description": "A document-level discount: `Kind` Amount (AED) or Percent (of the sub total); omit or use None for no discount."
      },
...
      "DocumentLineRequest": {
        "required": [
          "itemId",
          "description",
          "quantity",
          "unitPrice",
          "taxCategory"
        ],
        "type": "object",
        "properties": {
          "itemId": {
            "type": [
              "null",
              "integer"
            ],
            "format": "int32"
          },
          "description": {
            "type": "string"
          },
          "quantity": {
            "type": "number",
            "format": "double"
          },
          "unitPrice": {
            "type": "number",
            "format": "double"
          },
          "taxCategory": {
            "$ref": "#/components/schemas/TaxCategory"
          }
        },
        "description": "Validated by Task DocumentLines.ValidateAsync(IReadOnlyList&lt;DocumentLineRequest&gt;? lines, FatouraDbContext db, FieldValidator v, CancellationToken ct, bool allowInactiveItems = false) so all line errors are reported together."
      },
...
      "InvoiceDto": {
        "required": [
          "id",
          "number",
          "date",
          "clientId",
          "client",
          "company",
          "quotationId",
          "quotationNumber",
          "status",
          "voidReason",
          "voidedAt",
          "terms",
          "discount",
          "currency",
          "subTotal",
          "vatTotal",
          "total",
          "paidTotal",
          "creditedTotal",
          "balance",
          "createdById",
          "createdByName",
          "createdAt",
          "lines",
          "payments",
          "creditNotes"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "number": {
            "type": "string"
          },
          "date": {
            "type": "string",
            "format": "date"
          },
          "clientId": {
            "type": "integer",
            "format": "int32"
          },
          "client": {
            "$ref": "#/components/schemas/PartyDto"
          },
          "company": {
            "$ref": "#/components/schemas/CompanyDto"
          },
          "quotationId": {
            "type": [
              "null",
              "integer"
            ],
            "format": "int32"
          },
          "quotationNumber": {
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/InvoiceStatus"
          },
          "voidReason": {
            "type": "string"
          },
          "voidedAt": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "terms": {
            "$ref": "#/components/schemas/TermsDto"
          },
          "discount": {
            "$ref": "#/components/schemas/DocumentDiscountDto"
          },
          "currency": {
            "$ref": "#/components/schemas/DocumentCurrencyDto"
          },
          "subTotal": {
            "type": "number",
            "format": "double"
          },
          "vatTotal": {
            "type": "number",
            "format": "double"
          },
          "total": {
            "type": "number",
            "format": "double"
          },
          "paidTotal": {
            "type": "number",
            "format": "double"
          },
          "creditedTotal": {
            "type": "number",
            "format": "double"
          },
          "balance": {
            "type": "number",
            "format": "double"
          },
          "createdById": {
            "type": "string",
            "format": "uuid"
          },
          "createdByName": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "lines": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/InvoiceLineDto"
            }
          },
          "payments": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PaymentDto"
            }
          },
          "creditNotes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CreditNoteSummaryDto"
            }
          }
        }
      },
...
      "InvoiceResult": {
        "required": [
          "invoice",
          "warnings"
        ],
        "type": "object",
        "properties": {
          "invoice": {
            "$ref": "#/components/schemas/InvoiceDto"
          },
          "warnings": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/StockWarning"
            }
          }
        }
      },
      "InvoiceStatus": {
        "enum": [
          "Issued",
          "Void"
        ]
      },
...
      "PagedResultOfQuotationSummaryDto": {
        "required": [
          "items",
          "total",
          "page",
          "pageSize"
        ],
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/QuotationSummaryDto"
            }
          },
          "total": {
            "type": "integer",
            "format": "int32"
          },
          "page": {
            "type": "integer",
            "format": "int32"
          },
          "pageSize": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
...
      "QuotationDto": {
        "required": [
          "id",
          "number",
          "date",
          "validUntil",
          "isExpired",
          "clientId",
          "client",
          "status",
          "convertedInvoiceId",
          "convertedInvoiceNumber",
          "terms",
          "discount",
          "currency",
          "subTotal",
          "vatTotal",
          "total",
          "createdById",
          "createdByName",
          "createdAt",
          "lines"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "number": {
            "type": "string"
          },
          "date": {
            "type": "string",
            "format": "date"
          },
          "validUntil": {
            "type": "string",
            "format": "date"
          },
          "isExpired": {
            "type": "boolean"
          },
          "clientId": {
            "type": "integer",
            "format": "int32"
          },
          "client": {
            "$ref": "#/components/schemas/PartyDto"
          },
          "status": {
            "$ref": "#/components/schemas/QuotationStatus"
          },
          "convertedInvoiceId": {
            "type": [
              "null",
              "integer"
            ],
            "format": "int32"
          },
          "convertedInvoiceNumber": {
            "type": [
              "null",
              "string"
            ]
          },
          "terms": {
            "$ref": "#/components/schemas/TermsDto"
          },
          "discount": {
            "$ref": "#/components/schemas/DocumentDiscountDto"
          },
          "currency": {
            "$ref": "#/components/schemas/DocumentCurrencyDto"
          },
          "subTotal": {
            "type": "number",
            "format": "double"
          },
          "vatTotal": {
            "type": "number",
            "format": "double"
          },
          "total": {
            "type": "number",
            "format": "double"
          },
          "createdById": {
            "type": "string",
            "format": "uuid"
          },
          "createdByName": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "lines": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DocumentLineDto"
            }
          }
        }
      },
      "QuotationRequest": {
        "required": [
          "clientId",
          "date",
          "validUntil",
          "lines",
          "terms"
        ],
        "type": "object",
        "properties": {
          "clientId": {
            "type": "integer",
            "format": "int32"
          },
          "date": {
            "type": [
              "null",
              "string"
            ],
            "format": "date"
          },
          "validUntil": {
            "type": [
              "null",
              "string"
            ],
            "format": "date"
          },
          "lines": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DocumentLineRequest"
            }
          },
          "terms": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/TermsRequest"
              }
            ]
          },
          "discount": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/DocumentDiscountRequest"
              }
            ]
          },
          "currency": {
            "maxLength": 3,
            "type": [
              "null",
              "string"
            ]
          },
          "exchangeRate": {
            "type": [
              "null",
              "number"
            ],
            "format": "double"
          }
        }
      },
      "QuotationStatus": {
        "enum": [
          "Draft",
          "Sent",
          "Accepted",
          "Rejected",
          "Converted"
        ]
      },
      "QuotationStatusRequest": {
        "required": [
          "status"
        ],
        "type": "object",
        "properties": {
          "status": {
            "$ref": "#/components/schemas/QuotationStatus"
          }
        }
      },
      "QuotationSummaryDto": {
        "required": [
          "id",
          "number",
          "date",
          "validUntil",
          "isExpired",
          "clientId",
          "clientName",
          "status",
          "currency",
          "total",
          "convertedInvoiceId",
          "createdByName"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "number": {
            "type": "string"
          },
          "date": {
            "type": "string",
            "format": "date"
          },
          "validUntil": {
            "type": "string",
            "format": "date"
          },
          "isExpired": {
            "type": "boolean"
          },
          "clientId": {
            "type": "integer",
            "format": "int32"
          },
          "clientName": {
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/QuotationStatus"
          },
          "currency": {
            "type": "string"
          },
          "total": {
            "type": "number",
            "format": "double"
          },
          "convertedInvoiceId": {
            "type": [
              "null",
              "integer"
            ],
            "format": "int32"
          },
          "createdByName": {
            "type": "string"
          }
        }
      },
...
      "TaxCategory": {
        "enum": [
          "Standard",
          "ZeroRated",
          "Exempt"
        ]
      },
...
      "UpdateInvoiceRequest": {
        "required": [
          "clientId",
          "date",
          "lines",
          "terms"
        ],
        "type": "object",
        "properties": {
          "clientId": {
            "type": "integer",
            "format": "int32"
          },
          "date": {
            "type": "string",
            "format": "date"
          },
          "lines": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DocumentLineRequest"
            }
          },
          "terms": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/TermsRequest"
              }
            ]
          },
          "discount": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/DocumentDiscountRequest"
              }
            ]
          },
          "currency": {
            "maxLength": 3,
            "type": [
              "null",
              "string"
            ]
          },
          "exchangeRate": {
            "type": [
              "null",
              "number"
            ],
            "format": "double"
          }
        },
        "description": "Omitting the currency or rate keeps the invoice's own."
      },
...
    }
  },
...
}